valuable_lesson_learned
commit more often otherwise you might have to redo a finished project from scratch. 
the silver lining is that i got to become much more familiar with how using Python & Pandas with APIs for data analysis, manipulation and visualizations  work.
so that is the good thing.  i'm a silver lining kind of guy!